rework loading of history
On Jenkins instances with many jobs and runs the loading seems to cause
hanging jobs. I could see pipeline jobs finished but they exector on the
agent they run on was still in use. A possible cause might be that the
complete history was loaded whenever an agent was used the first time.
As loading looped over all runs this unnecessarily made it loop over all
runs several times (which can be quite time consuming on a Jenkins with
> 150 agents and > 50000 runs).
THis is now avoided and history is only loaded the very first time an
extended history page is opended. Though this means that the history
will only contain those runs that started running on an executor since
Jenkins started. But from that point on the history should be complete.
There might be corner cases where history is missing (e.g. an agent is
deleted and then recreated).