fix runs are added to all nodes on completion
the updateResult was wrongly adding a run to the list when there was no
entry for the run. But here adding the run is wrong as we should only
update the result. When the run is not found it means the run was not
happening on that node.

Also write the file only when it has really changed