Reduce I/O
Add an action to pipeline runs that keeps track of the used agents. When
updating the result, use this to only update the files that have really
been used by the run. This avoids to read all the files of unused
agents.
Similarly AbstractBuilds just use a single agent so just use this
instead.

Use a FileWriter instead of wrapping FileOutputStream with
an OutputStreamWrite
Use Files.readAllLines to read an indexfile